+++ packages/video_player/video_player_avfoundation/CHANGELOG.md
@@ -1,5 +1,7 @@
-## NEXT
+## 2.6.2
 
+* Adds possibility to play videos at more than 30 FPS.
+* Fixes playing state not updating in some paths.
 * Updates minimum supported SDK version to Flutter 3.19/Dart 3.3.
 
 ## 2.6.1